propose edit data duplicate

+++ src/main/java/com/clevel/selos/businesscontrol/CreditFacProposeControl.java
@@ -259,12 +259,10 @@ public class CreditFacProposeControl extends BusinessControl {
                                     if (productFormula.getExposureMethod() == ExposureMethod.NOT_CALCULATE.value()) { //ไม่คำนวณ
                                         log.debug("NOT_CALCULATE :: productFormula.getExposureMethod() :: {}", productFormula.getExposureMethod());
                                         sumTotalPropose = sumTotalPropose.add(BigDecimal.ZERO);
-                                    }
-                                    else if (productFormula.getExposureMethod() == ExposureMethod.LIMIT.value()) { //limit
+                                    } else if (productFormula.getExposureMethod() == ExposureMethod.LIMIT.value()) { //limit
                                         log.debug("LIMIT :: productFormula.getExposureMethod() :: {}", productFormula.getExposureMethod());
                                         sumTotalPropose = sumTotalPropose.add(newCreditDetailView.getLimit());
-                                    }
-                                    else if (productFormula.getExposureMethod() == ExposureMethod.PCE_LIMIT.value()) {    //limit * %PCE
+                                    } else if (productFormula.getExposureMethod() == ExposureMethod.PCE_LIMIT.value()) {    //limit * %PCE
                                         log.debug("PCE_LIMIT :: productFormula.getExposureMethod() :: {}", productFormula.getExposureMethod());
                                         sumTotalPropose = sumTotalPropose.add(Util.multiply(newCreditDetailView.getLimit(), newCreditDetailView.getPCEPercent()));
                                     }
@@ -274,18 +272,15 @@ public class CreditFacProposeControl extends BusinessControl {
                                     if (productFormula.getDbrCalculate() == 1) {// No
                                         log.info("NO calculate :: productFormula.getDbrCalculate() :: {}", productFormula.getDbrCalculate());
                                         sumTotalNonLoanDbr = BigDecimal.ZERO;
-                                    }
-                                    else if (productFormula.getDbrCalculate() == 2) {// Yes
+                                    } else if (productFormula.getDbrCalculate() == 2) {// Yes
                                         log.debug("YES :: productFormula.getDbrCalculate() :: {}", productFormula.getDbrCalculate());
                                         if (productFormula.getDbrMethod() == DBRMethod.NOT_CALCULATE.value()) {// not calculate
                                             log.debug("NOT_CALCULATE :: productFormula.getDbrMethod() :: {}", productFormula.getDbrMethod());
                                             sumTotalLoanDbr = sumTotalLoanDbr.add(BigDecimal.ZERO);
-                                        }
-                                        else if (productFormula.getDbrMethod() == DBRMethod.INSTALLMENT.value()) { //Installment
+                                        } else if (productFormula.getDbrMethod() == DBRMethod.INSTALLMENT.value()) { //Installment
                                             log.debug("INSTALLMENT :: productFormula.getDbrMethod() :: {}", productFormula.getDbrMethod());
                                             sumTotalLoanDbr = sumTotalLoanDbr.add(newCreditDetailView.getInstallment());
-                                        }
-                                        else if (productFormula.getDbrMethod() == DBRMethod.INT_YEAR.value()) { //(Limit*((อัตราดอกเบี้ย+ Spread)/100))/12
+                                        } else if (productFormula.getDbrMethod() == DBRMethod.INT_YEAR.value()) { //(Limit*((อัตราดอกเบี้ย+ Spread)/100))/12
                                             log.debug("INT_YEAR :: productFormula.getDbrMethod() :: {}, productFormula.getDbrSpread() :::{}", productFormula.getDbrMethod(), productFormula.getDbrSpread());
                                             sumTotalLoanDbr = sumTotalLoanDbr.add(calTotalProposeLoanDBRForIntYear(newCreditDetailView, productFormula.getDbrSpread()));
                                         }
@@ -322,10 +317,11 @@ public class CreditFacProposeControl extends BusinessControl {
     public BigDecimal calTotalProposeLoanDBRForIntYear(NewCreditDetailView newCreditDetailView, BigDecimal dbrSpread) {
         log.debug("calTotalProposeLoanDBRForIntYear start :: newCreditDetailView and  dbrSpread ::{}", newCreditDetailView, dbrSpread);
         log.info("limit :: {}", newCreditDetailView.getLimit());
-        log.info("newCreditTierDetailViews.size :: {}", newCreditDetailView.getNewCreditTierDetailViewList().size());
 
         BigDecimal sumTotalLoanDbr = BigDecimal.ZERO;
         if (newCreditDetailView.getNewCreditTierDetailViewList() != null) {
+            log.info("newCreditTierDetailViews.size :: {}", newCreditDetailView.getNewCreditTierDetailViewList().size());
+
             BigDecimal oneHundred = new BigDecimal("100");
             BigDecimal twelve = new BigDecimal("12");
             BigDecimal sum = BigDecimal.ZERO;
@@ -389,8 +385,7 @@ public class CreditFacProposeControl extends BusinessControl {
                             // Count core asset and none core asset
                             if (PotentialCollateralValue.CORE_ASSET.id() == potentialCollateral.getId()) {
                                 totalNumOfCoreAsset = Util.add(totalNumOfCoreAsset, BigDecimal.ONE);
-                            }
-                            else if (PotentialCollateralValue.NONE_CORE_ASSET.id() == potentialCollateral.getId()) {
+                            } else if (PotentialCollateralValue.NONE_CORE_ASSET.id() == potentialCollateral.getId()) {
                                 totalNumOfNonCoreAsset = Util.add(totalNumOfNonCoreAsset, BigDecimal.ONE);
                             }
 
@@ -419,11 +414,9 @@ public class CreditFacProposeControl extends BusinessControl {
                         if (customerEntity != null) {
                             if (GuarantorCategory.INDIVIDUAL.value() == customerEntity.getId()) {
                                 totalIndiGuaranteeAmt = Util.add(totalIndiGuaranteeAmt, guarantorDetailView.getTotalLimitGuaranteeAmount());
-                            }
-                            else if (GuarantorCategory.JURISTIC.value() == customerEntity.getId()) {
+                            } else if (GuarantorCategory.JURISTIC.value() == customerEntity.getId()) {
                                 totalJuriGuaranteeAmt = Util.add(totalJuriGuaranteeAmt, guarantorDetailView.getTotalLimitGuaranteeAmount());
-                            }
-                            else if (GuarantorCategory.TCG.value() == customerEntity.getId()) {
+                            } else if (GuarantorCategory.TCG.value() == customerEntity.getId()) {
                                 totalTCGGuaranteeAmt = Util.add(totalTCGGuaranteeAmt, guarantorDetailView.getTotalLimitGuaranteeAmount());
                             }
                         }
@@ -815,8 +808,10 @@ public class CreditFacProposeControl extends BusinessControl {
         if (Util.safetyList(newCreditFacilityView.getNewGuarantorDetailViewList()).size() > 0) {
 
             List<NewGuarantorDetail> tmpNewGuarantorList = newGuarantorDetailDAO.findNewGuarantorByNewCreditFacility(newCreditFacility);
-            for(NewGuarantorDetail newGuarantorDetail : tmpNewGuarantorList){
-                newGuarantorRelationDAO.delete(newGuarantorDetail.getNewGuarantorCreditList());
+            for (NewGuarantorDetail newGuarantorDetail : tmpNewGuarantorList) {
+                if (newGuarantorDetail.getNewGuarantorCreditList() != null) {
+                    newGuarantorRelationDAO.delete(newGuarantorDetail.getNewGuarantorCreditList());
+                }
                 newGuarantorDetail.setNewGuarantorCreditList(Collections.<NewGuarantorCredit>emptyList());
                 newGuarantorDetailDAO.persist(newGuarantorDetail);
             }
@@ -828,28 +823,30 @@ public class CreditFacProposeControl extends BusinessControl {
 
         }
 
-        //--- Need to Delete SubMortgage from CollateralSubMortgages before Insert new
-//        List<NewCollateralSubMortgage> newCollateralSubMortgages = newSubCollMortgageDAO.getListByWorkCase(workCase, ProposeType.P);
-//        log.debug("before :: newCollateralSubMortgages :: size :: {}", newCollateralSubMortgages.size());
-//        newSubCollMortgageDAO.delete(newCollateralSubMortgages);
-//        log.debug("after :: newCollateralSubMortgages :: size :: {}", newCollateralSubMortgages.size());
-        //--- Need to Delete SubOwner from CollateralSubOwner before Insert new
-//        List<NewCollateralSubOwner> newCollateralSubOwnerList = newCollateralSubOwnerDAO.getListByWorkCase(workCase, ProposeType.P);
-//        log.debug("before :: newCollateralSubOwnerList :: size :: {}", newCollateralSubOwnerList.size());
-//        newCollateralSubOwnerDAO.delete(newCollateralSubOwnerList);
-//        log.debug("before :: newCollateralSubOwnerList :: size :: {}", newCollateralSubOwnerList.size());
-        //--- Need to Delete SubOwner from newCollateralSubRelatedList before Insert new
-//        List<NewCollateralSubRelated> newCollateralSubRelatedList = newCollateralSubRelatedDAO.getListByWorkCase(workCase, ProposeType.P);
-//        log.debug("before :: newCollateralSubRelatedList :: size :: {}",newCollateralSubRelatedList.size());
-//        newCollateralSubRelatedDAO.delete(newCollateralSubRelatedList);
-//        log.debug("before :: newCollateralSubRelatedList :: size :: {}",newCollateralSubRelatedList.size());
-
-        if (Util.safetyList(newCreditFacilityView.getNewCollateralViewList()).size() > 0){
+        if (Util.safetyList(newCreditFacilityView.getNewCollateralViewList()).size() > 0) {
 
             List<NewCollateral> tmpNewCollateralList = newCollateralDAO.findNewCollateralByNewCreditFacility(newCreditFacility);
-            for(NewCollateral newCollateral : tmpNewCollateralList){
-                newCollateralRelationDAO.delete(newCollateral.getNewCollateralCreditList());
-                newCollateral.setNewCollateralCreditList(Collections.<NewCollateralCredit>emptyList());
+            for (NewCollateral newCollateral : tmpNewCollateralList) {
+                if (newCollateral.getNewCollateralCreditList() != null) {
+                    newCollateralRelationDAO.delete(newCollateral.getNewCollateralCreditList());
+                    newCollateral.setNewCollateralCreditList(Collections.<NewCollateralCredit>emptyList());
+                }
+//                List<NewCollateralHead> newCollateralHeadList = newCollateralHeadDetailDAO.findByNewCollateral(newCollateral);
+                List<NewCollateralHead> newCollateralHeadList = newCollateral.getNewCollateralHeadList();
+                for(NewCollateralHead newCollateralHead:newCollateralHeadList)
+                {
+//                    List<NewCollateralSub> newCollateralSubList = newCollateralSubDetailDAO.findByNewCollateralHead(newCollateralHead);
+                    List<NewCollateralSub> newCollateralSubList = newCollateralHead.getNewCollateralSubList();
+                    for(NewCollateralSub newCollateralSub:newCollateralSubList)
+                    {
+                        newSubCollMortgageDAO.delete(newCollateralSub.getNewCollateralSubMortgageList());
+                        newCollateralSubOwnerDAO.delete(newCollateralSub.getNewCollateralSubOwnerList());
+                        newCollateralSub.setNewCollateralSubMortgageList(Collections.<NewCollateralSubMortgage>emptyList());
+                        newCollateralSub.setNewCollateralSubOwnerList(Collections.<NewCollateralSubOwner>emptyList());
+//                        newCollateralSubDetailDAO.persist(newCollateralSub);
+                    }
+//                    newCollateralHeadDetailDAO.persist(newCollateralHead);
+                }
                 newCollateralDAO.persist(newCollateral);
             }
 
@@ -906,24 +903,24 @@ public class CreditFacProposeControl extends BusinessControl {
 //            }
 
 
-            for (PricingFee pricingFee : standardPricingResponse.getPricingFeeList()){
+            for (PricingFee pricingFee : standardPricingResponse.getPricingFeeList()) {
                 FeeDetailView feeDetailView = feeTransform.transformToView(pricingFee);
                 // find productProgram
-                ProductProgramView productProgramView = productTransform.transformToView(productProgramDAO.findById((int)feeDetailView.getCreditDetailViewId()));
+                ProductProgramView productProgramView = productTransform.transformToView(productProgramDAO.findById((int) feeDetailView.getCreditDetailViewId()));
                 newFeeDetailView.setProductProgram(productProgramView.getDescription());
                 if (feeDetailView.getFeeTypeView().getId() == 9) {//type=9,(Front-End-Fee)
                     newFeeDetailView.setStandardFrontEndFee(feeDetailView);
-                }else if (feeDetailView.getFeeTypeView().getId() == 15) { //type=15,(Prepayment Fee)
+                } else if (feeDetailView.getFeeTypeView().getId() == 15) { //type=15,(Prepayment Fee)
                     newFeeDetailView.setPrepaymentFee(feeDetailView);
-                }else if (feeDetailView.getFeeTypeView().getId() == 20) {//type=20,(CancellationFee)
+                } else if (feeDetailView.getFeeTypeView().getId() == 20) {//type=20,(CancellationFee)
                     newFeeDetailView.setCancellationFee(feeDetailView);
-                }else if (feeDetailView.getFeeTypeView().getId() == 21) { //type=21,(ExtensionFee)
+                } else if (feeDetailView.getFeeTypeView().getId() == 21) { //type=21,(ExtensionFee)
                     newFeeDetailView.setExtensionFee(feeDetailView);
-                }else  if(feeDetailView.getFeeTypeView().getId()==22){//type=22,(CommitmentFee)
+                } else if (feeDetailView.getFeeTypeView().getId() == 22) {//type=22,(CommitmentFee)
                     newFeeDetailView.setCommitmentFee(feeDetailView);
                 }
 
-                log.debug("FeePaymentMethodView():::: {}",feeDetailView.getFeePaymentMethodView().getBrmsCode());
+                log.debug("FeePaymentMethodView():::: {}", feeDetailView.getFeePaymentMethodView().getBrmsCode());
                 newFeeDetailViewList.add(newFeeDetailView);
             }
